Modify _read_submissions to handle scenario when no submission file exists

+++ xdyna/da_meta.py
@@ -674,18 +674,28 @@ class _DAMetaData:
     # This will overwrite the value associated to submission_id in self._submissions with val.
     def _read_submissions(self):
         if self._use_files and not self._read_only:
-            with ProtectFile(self.submissions_file, 'r+', wait=0.005) as pf:
+            val = {}
+            if self.submissions_file.exists():
                 try:
-                    val = json.load(pf)
+                    with ProtectFile(self.submissions_file, 'r+', wait=0.005) as pf:
+                        val = json.load(pf)
                 except json.decoder.JSONDecodeError:
                     # This file is expected to be easily corrupted. If this happen, then it need to be rebuild
                     # using backups file from submissions_dir
                     val = self.rebuild_submission_from_backup()
+                    with ProtectFile(self.submissions_file, 'r+', wait=0.005) as pf:
+                        pf.truncate(0)  # Delete file contents (to avoid appending)
+                        pf.seek(0)      # Move file pointer to start of file
+                        json.dump(val, pf, indent=2, sort_keys=False)
+                    #setattr(self, '_submissions', val['submissions'] )
+            elif self.submission_backup('0').exists():
+                val = self.rebuild_submission_from_backup()
+                with ProtectFile(self.submissions_file, 'x+', wait=0.005) as pf:
                     pf.truncate(0)  # Delete file contents (to avoid appending)
                     pf.seek(0)      # Move file pointer to start of file
                     json.dump(val, pf, indent=2, sort_keys=False)
-                #setattr(self, '_submissions', val['submissions'] )
-                setattr(self, '_submissions', val )
+                
+            setattr(self, '_submissions', val )
 
     # Allowed on parallel process (but only if each process updates only the log attached to its unique ID).
     # This will overwrite the value associated to submission_id in self._submissions with val.
